requirements.txt
## Python 3.6.13

absl-py==0.15.0
astor==0.8.1
astunparse==1.6.3
backports.weakref==1.0rc1
bleach==1.5.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
clang==5.0
cycler==0.11.0
dataclasses==0.8
flatbuffers==1.12
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==2.10.0
html5lib==0.9999999
idna==3.7
importlib-metadata==4.8.3
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==2.2.0
matplotlib==3.3.4
numpy==1.18.5
oauthlib==3.2.2
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9
six==1.15.0
tensorboard==1.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==1.9.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.19
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0

neat-python==0.92